Update dofus version (2.60.1.4)

+++ BotDofus/src/Protocol/Messages/Data/game/guild/tax/GuildFightTakePlaceRequestMessage.h
@@ -16,7 +16,7 @@ public:
   void deserializeAsyncAs_GuildFightTakePlaceRequestMessage(FuncTree tree);
   GuildFightTakePlaceRequestMessage();
 
-  int replacedCharacterId;
+  double replacedCharacterId;
 
 private:
   void _replacedCharacterIdFunc(Reader *input);
